Refactor:  web_crawling_retrieval starterkit structure adjusted
