+++ web_crawled_data_retriever/README.md
@@ -55,7 +55,7 @@ Integrate your LLM deployed on SambaStudio with this AI starter kit in two simpl
 ```
 git clone https://github.com/sambanova/ai-starter-kit.git
 ```
-2. Update API information for the SambaNova LLM and, optionally, the vector database.  These are represented as configurable variables in the environment variables file in sn-ai-starter-kit/web_crawled_data_retriever/export.env. For example, an endpoint with the URL
+2. Update API information for the SambaNova LLM and, optionally, the vector database.  These are represented as configurable variables in the environment variables file in sn-ai-starter-kit/export.env. For example, an endpoint with the URL
 "https://api-stage.sambanova.net/api/predict/nlp/12345678-9abc-def0-1234-56789abcdef0/456789ab-cdef-0123-4567-89abcdef0123"
 would be entered in the config file (with no spaces) as:
 ```
@@ -70,18 +70,19 @@ VECTOR_DB_URL=http://host.docker.internal:6333
 cd ai_starter_kit/web_crawled_data_retriever
 python3 -m venv web_crawling_env
 source web_crawling_env/bin/activate
-pip install -r requirements.txt
+pip install -r web_crawled_data_retriever/requirements.txt
 ```
 
 ## Deploy the starter kit
 To run the demo, run the following commands:
 ```
-sh run.sh
+cd web_crawled_data_retriever
+sh streamlit/run.sh
 ```
 
 After deploying the starter kit you should see the following application user interface
 
-![capture of web_crawled_retriever_demo](./web_crawled_app.png)
+![capture of web_crawled_retriever_demo](./docs/web_crawled_app.png)
 
 ## Starterkit usage 
 
@@ -159,23 +160,25 @@ The provided example template can be further customized based on the use case.
 
 **website scraping** Different packages are available to crawl and extract out of websites. In the demo app it is implemented the [AsyncHtmlLoader](https://python.langchain.com/docs/integrations/document_loaders/async_html), langchain also includes a cople of [HTML loaders](https://python.langchain.com/docs/modules/data_connection/document_loaders/html) that can be used.
 This modification can be done in the following location:
-```
-file: [app.py](app.py)
-function: load_htmls
-```
+
+> file: [src/web_crawling_retriever.py](src/web_crawling_retriever.py)
+>```
+>function: load_htmls
+>```
 
 **Iterative web crawling:** for each provided site after the scraping, all the referenced links are saved and filtered using [beautifulSoup](https://www.crummy.com/software/BeautifulSoup/) package, then the web crawling method iterates 'n' times scraping this sites and finding referenced links, this depth is limited to 2 maximum depth, but you can modify this limit, and the behavior of the web crawling in the following location:
-```
-file: [app.py](app.py)
-function: web_crawl
-```
+> file: [src/web_crawling_retriever.py](src/web_crawling_retriever.py)
+>```
+>function: web_crawl
+>```
+
 > Be cautious as the crawling depth increases, leading to exponential growth in the number of processed sites. Consider resource implications for efficient workflow performance.
 
 **Document transformations** Depending on the loader used for scraping the sites, you may want or not to use some transformation method to clean up the downloaded documents. , this could be done in the following location:
-```
-file: [app.py](app.py)
-function: clean_docs
-```
+> file: [src/web_crawling_retriever.py](src/web_crawling_retriever.py)
+>```
+>function: clean_docs
+>```
 *[LangChain](https://python.langchain.com/docs/integrations/document_transformers) provides several document transformers that can be used with you data*
 
 ## Split Data
@@ -195,10 +198,12 @@ chunk_overlap=20
 
 
 This modification can be done in the following location:
-```
-file: app.py
-function: get_text_chunks
-```
+> file: [src/web_crawling_retriever.py](src/web_crawling_retriever.py)
+>```
+>Atributes: 
+> - LLM_MAX_TOKENS_TO_GENERATE = 500
+> - CHUNK_SIZE = 1200
+>```
 
 
 ## Embed data
@@ -206,10 +211,10 @@ function: get_text_chunks
 There are several open-source embedding models available on HuggingFace. [This leaderboard](https://huggingface.co/spaces/mteb/leaderboard) ranks these models based on the Massive Text Embedding Benchmark (MTEB). A number of these models are available on SambaStudio and can be further fine-tuned on specific datasets to improve performance.
 
 This modification can be done in the following location:
-```
-file: app.py
-function: get_vectorstore
-```
+> file: [../vectordb/vector_db.py](../vectordb/vector_db.py)
+>```
+> function: load_embedding_model
+>```
 
 
 ## Store embeddings
@@ -217,10 +222,10 @@ function: get_vectorstore
 The template can be customized to use different vector databases to store the embeddings generated by the embedding model. The [LangChain vector stores documentation](https://python.langchain.com/docs/integrations/vectorstores) provides a broad collection of vector stores that can be easily integrated.
 
 This modification can be done in the following location:
-```
-file: app.py
-function: get_vectorstore
-```
+> file: [../vectordb/vector_db.py](../vectordb/vector_db.py)
+>```
+> function: create_vector_store
+>```
 
 
 ## Retrieval
@@ -229,10 +234,10 @@ Similar to the vector stores, a wide collection of retriever options is also ava
 
 
 This modification can be done in the following location:
-```
-file: app.py
-function: get_conversation_chain
-```
+> file: [src/web_crawling_retriever.py](src/web_crawling_retriever.py)
+>```
+>function: retrieval_qa_chain
+>```
 
 
 ## Large language model (LLM)
@@ -243,21 +248,22 @@ The template uses the SN LLM model, which can be further fine-tuned to improve r
 ### Prompt engineering
 
 Finally, prompting has a significant effect on the quality of LLM responses. Prompts can be further customized to improve the overall quality of the responses from the LLMs. For example, in the given template, the following prompt was used to generate a response from the LLM, where ```question``` is the user query and ```context``` are the documents retrieved by the retriever.
-```python
-custom_prompt_template = """Use the following pieces of context to answer the question at the end. If the answer to the question cannot be extracted from given CONTEXT than say I do not have information regarding this.
-{context}
-
-Question: {question}
-Helpful Answer:"""
-CUSTOMPROMPT = PromptTemplate(
-template=custom_prompt_template, input_variables=["context", "question"]
+```yaml
+template: |
+          <s>[INST] <<SYS>>\nUse the following pieces of context to answer the question at the end.
+          If the answer is not in context for answering, say that you don't know, don't try to make up an answer or provide an answer not extracted from provided context.
+          Cross check if the answer is contained in provided context. If not than say \"I do not have information regarding this.\"\n
+          context
+          {context}
+          end of context
+          <</SYS>>/n
+          Question: {question}
+          Helpful Answer: [/INST]
 )
 ```
 This modification can be done in the following location:
-```
-file: app.py
-function: get_conversation_chain
-```
+> file: [prompts/llama7b-web_crwling_data_retriever.yaml](prompts/llama7b-web_crwling_data_retriever.yaml)
+
 > *Learn more about [Prompt engineering](https://www.promptingguide.ai/)*
 
 # Third-party tools and data sources
