+++ web_crawled_data_retriever/prompts/llama7b-web_crwling_data_retriever.yaml
@@ -1,0 +1,20 @@
+_type: prompt
+input_types: {}
+input_variables:
+  - context
+  - question
+name: null
+output_parser: null
+partial_variables: {}
+template: |
+          <s>[INST] <<SYS>>\nUse the following pieces of context to answer the question at the end.
+          If the answer is not in context for answering, say that you don't know, don't try to make up an answer or provide an answer not extracted from provided context.
+          Cross check if the answer is contained in provided context. If not than say \"I do not have information regarding this.\"\n
+          context
+          {context}
+          end of context
+          <</SYS>>/n
+          Question: {question}
+          Helpful Answer: [/INST]
+template_format: f-string
+validate_template: false
