--- web_crawled_data_retriever/run.sh
@@ -1,2 +1,0 @@
-set -o allexport && source export.env && set +o allexport
-streamlit run app.py
